refactor(framework): split isomorphic chat execution
Refactor isomorphic tool and session orchestration into focused modules so streaming, handoff, and state transitions are easier to reason about and maintain.

Add client-only isomorphic tool support and centralize conversation/history sync for elicit + tool-call flows to prevent continuation mismatches and reduce "No tool call found" errors.
